+++ src/main/java/longah/handler/Logging.java
@@ -1,4 +1,4 @@
-package longah.util;
+package longah.handler;
 
 import java.util.logging.Logger;
 import java.io.IOException;
@@ -7,20 +7,31 @@ import java.util.logging.SimpleFormatter;
 import java.util.logging.Level;
 
 public class Logging {
-    private static final Logger LongAhLogger = Logger.getLogger("LongAh");
+    private static Logger LongAhLogger = null;
 
     /**
      * Constructs a new Logging instance.
      * Creates a log file to store log data.
      */
-    public Logging() {
+    private Logging() {
         try {
+            LongAhLogger = Logger.getLogger("LongAh");
             FileHandler handler = new FileHandler("./log/LongAh.log");
             handler.setFormatter(new SimpleFormatter());
             LongAhLogger.addHandler(handler);
             LongAhLogger.setUseParentHandlers(false);
         } catch (IOException e) {
             LongAhLogger.log(Level.WARNING, "Log data may not be saved due to permission.");
+        }
+    }
+
+    /**
+     * Initializes the logger.
+     * Enables singleton pattern for the logger.
+     */
+    public static void initLogger() {
+        if (LongAhLogger == null) {
+            new Logging();
         }
     }
